+++ spec/swagger_v2/api_documentation_spec.rb
@@ -18,25 +18,9 @@ describe 'API with additional options' do
   end
 
   it 'documents api' do
-    expect(subject).to eq(
-      [
-        { description: 'Swagger compatible API description' },
-        {
-          description: 'Swagger compatible API description for specific API',
-          params: {
-            'locale' => {
-              desc: 'Locale of API documentation',
-              required: false,
-              type: 'Symbol'
-            },
-            'name' => {
-              desc: 'Resource name of mounted API',
-              required: true,
-              type: 'String'
-            }
-          }
-        }
-      ]
-    )
+    expect(subject.pluck(:description)).to match_array [
+      'Swagger compatible API description',
+      'Swagger compatible API description for specific API'
+    ]
   end
 end